Refactor useDarkMode.js and page.js to update dark mode functionality
